requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.25.2
opencv-python==4.8.1.78
tensorflow-macos>=2.9.0
tensorflow-metal>=0.7.0
mediapipe==0.10.7
python-multipart==0.0.6
pillow==10.1.0
jinja2==3.1.2
aiofiles==23.2.1
onnxruntime==1.16.0
scikit-learn==1.3.2
matplotlib==3.8.1
tf2onnx>=1.13.0
websockets>=11.0.3
starlette>=0.27.0
pydantic>=2.0.0
typing-extensions>=4.7.0
# Optional dependencies
h5py>=3.9.0
# For testing/development
pytest>=7.4.0
requests>=2.31.0
 